feat: Ajoute une interface de tracking globale
