+++ src/lib/statistics-service/matomo.ts
@@ -1,5 +1,6 @@
 import { skipSendStatistics } from "./shared.js"
 import { EventCategory } from "@lib/enums/event-category.js"
+import tracker from "@/plugins/tracker.js"
 
 const isProduction = process.env.NODE_ENV === "production"
 
@@ -10,25 +11,11 @@ export interface MatomoEvent {
   value?: string
 }
 
-export interface Matomo {
-  trackEvent: (
-    category: EventCategory,
-    action: string,
-    label: string,
-    value?: string
-  ) => void
-  getVisitorId: () => string
-}
-
-function skipSendEventToMatomo(matomo: Matomo | undefined): boolean {
-  return skipSendStatistics() || matomo === undefined
-}
-
-export function sendEventToMatomo(event: MatomoEvent, matomo: Matomo): void {
-  if (skipSendEventToMatomo(matomo)) {
+export function sendEventToMatomo(event: MatomoEvent): void {
+  if (skipSendStatistics()) {
     !isProduction && console.debug("Skip sending event to Matomo", event)
     return
   }
 
-  matomo.trackEvent(event.category, event.action, event.label, event.value)
+  tracker.trackEvent(event.category, event.action, event.label, event.value)
 }
